🔵 rename "operator.ts" into "operatorFeatures.ts"
